orgchart-api-data/src/main/java/com.nexient.orgchart/data/entity/JobTitle.java
package com.nexient.orgchart.data.entity;

import org.hibernate.validator.constraints.NotEmpty;

import javax.persistence.CascadeType;
import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.OneToMany;
import javax.persistence.Table;
import javax.validation.constraints.NotNull;
import java.util.HashSet;
import java.util.Set;

@Entity
@Table(name = "JOB_TITLE")
public class JobTitle extends BaseEntity {

	@Column(name = "NAME", unique = true, nullable = false)
	@NotNull
	@NotEmpty
	private String name;

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}
}